Feature: spinner for blocking the browser and means to display it
To block the browser just call foswiki.ModacSkin.blockUI(), to release
it again call foswiki.ModacSkin.unblockUI().

For custom blocks call <some jQuery object>.block(foswiki.ModacSkin.blockDefaultOptions)

In order to customize overwrite div.ajaxspinner

Unfortunately jQuery can not calculate the dimensions of the overlay
correctly when placing it (width: auto, however the element is not yet in the dom), thus
the spinner will be displayed a bit to the left of the center.

Ticket: #3359