metacomments: remove unused CSS class that autolinks
During previous development a CSS class was added (but not used) that
happens to be a WikiWord. This causes rendering issues when autolinking
is enabled. This is not our default so can only possibly cause issues in
extremely non-standard installs. For this reason I am not listing this
change in the release notes.

Ticket: none
Release-Note: none