Fix facets removing "extrafilter"
When a facet is selected, ALL other filters on the same field are
removed, so if the solr::extrafilter started with "-type:..." all
extrafilters were removed, as soon as a "type" filter was selected.

As a workaround, I'm starting this filter with a field that is rather
unlikely to be faceted on.

Ticket: 9312

Release-Note(fix, user) {
  Fixing a bug, where selecting a facet removed default filters.
[...]
  When selecting a "type" facet, other default filters (eg. filtering
the System web) got removed. All default filters should persist now.
}